split out snql_api join tests
